# api/openapi.yaml
openapi: 3.1.0
info:
  title: yamo
  version: "1.0.0"
  summary: Catalogue, search and edit the metadata of a large music library.
  description: |
    yamo keeps a catalogue of a music library and edits the tags in the files
    themselves. This API is the only way anything reaches those files: the
    terminal interface and the command line are both clients of it.

    ## Working with a large library

    The library is expected to run to six figures, and the work is gradual —
    find a few wrong things, fix them, carry on. Three parts of the design
    follow from that.

    **Everything is paged.** `GET /tracks` returns a window and a total. There
    is no endpoint that returns the whole library.

    ## Paging

    Paging is offset-based. `limit` and `offset` go in the query string; the
    response is an envelope of `items`, `total`, `limit` and `offset`.

    ```
    GET /v1/tracks?q=artist:elvis&sort=album,track&limit=50&offset=100
    → { "items": [ … 50 tracks … ], "total": 2400, "limit": 50, "offset": 100 }
    ```

    - `total` is the number of matches **before** paging, so it is what a
      result count or a scrollbar should be sized from.
    - `limit` and `offset` are echoed back as applied. A `limit` above the
      maximum is capped rather than rejected, so read the response rather than
      assuming the request was honoured.
    - You have reached the end when `offset + len(items) >= total`, or when
...
      as offset 0.

...

...
    not a page of a larger set.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /jobs/{id}/events:
    get:
      tags: [jobs]
      summary: Stream a job's progress
      description: A `text/event-stream` of `Event` objects for one job.
      operationId: streamJobEvents
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required whenever the server is bound to anything other than loopback.
        The token is generated on first run and printed once.
...
  responses:
    NotFound:
      description: No such resource.
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: A message suitable for showing to a person. }
        code:
          type: string
          description: A stable identifier for the failure.
          enum: [not_found, bad_request, conflict, exists, count_mismatch, scan_running, unwritable, unauthorized, internal]
        jobId: { type: string, description: "For scan_running, the scan already in progress." }
        expected: { type: integer, description: "For count_mismatch, the count the client sent." }
        actual: { type: integer, description: "For count_mismatch, the count the server found." }
...